Howie/enable theming improved automation (#5)
* Enable comprehensive theming feature

This simply enables the comprehensive theming feature and places the theme files in the default theme dir.

* Add empty variable to demo role

Ansible will fail a task if a variable isn't present even if the task is skipped. This adds an empty list for demo_test_and_staff_users so that the play "create some test users" in the demo role won't fail if the task which sets this fact is skipped earlier in the role.